requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.23.0
autopy==4.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coloredlogs==14.3
cycler==0.11.0
docopt==0.6.2
Eel==0.14.0
ffmpeg==1.4
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.34.4
Frozen-Flask==0.18
frozenlist==1.3.3
fs==2.4.16
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
gunicorn==20.1.0
humanfriendly==10.0
idna==3.3
importlib-metadata==6.0.0
instalooter==2.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.2
mediapipe==0.8.10.1
mouse==0.7.1
MouseInfo==0.1.3
multidict==6.0.4
mutagen==1.45.1
mysql-connector-python==8.0.29
mysql-to-sqlite3==1.4.14
numpy==1.23.1
openai==0.26.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pafy==0.5.5
pefile==2022.5.30
Pillow==9.2.0
pipreqs==0.4.11
protobuf==3.20.1
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.1.2
PyGetWindow==0.0.9
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.28
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-slugify==5.0.2
python-telegram-bot==13.13
pytimeparse==1.1.8
pytube==12.1.0
pytweening==1.0.4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sockets==1.0.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tenacity==6.3.1
text-unidecode==1.3
tornado==6.2
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
verboselogs==1.7
waitress==2.1.2
Werkzeug==2.2.2
whichcraft==0.6.1
win10toast==0.9
yarg==0.1.9
yarl==1.8.2
youtube-dl==2021.12.17
youtube-transcript-api==0.4.4
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.4.0
